audio: volume is float

+++ src/engine/audio/audio.h
@@ -42,10 +42,10 @@ audio_asset_kind_t audio_asset_get_kind(int asset_handle);
 /* ------------------------------------------------------------------------- */
 
 /* returns voice handle */
-int  audio_play(int asset_handle, int volume_percent, float speed, int loop);
+int  audio_play(int asset_handle, float volume_percent, float speed, int loop);
 
 int  audio_play_ex(int asset_handle,
-                   int volume_percent,
+                   float volume_percent,
                    float speed,
                    int loop,
                    audio_voice_callback_t on_started,
@@ -56,8 +56,8 @@ void audio_voice_stop(int voice_handle);
 void audio_voice_pause(int voice_handle);
 void audio_voice_resume(int voice_handle);
 
-void audio_voice_set_volume(int voice_handle, int percent);
-void audio_voice_set_channel_volume(int voice_handle, int left_percent, int right_percent);
+void audio_voice_set_volume(int voice_handle, float percent);
+void audio_voice_set_channel_volume(int voice_handle, float left_percent, float right_percent);
 void audio_voice_set_pan(int voice_handle, float pan);
 void audio_voice_set_speed(int voice_handle, float speed);
 void audio_voice_set_callbacks(int voice_handle,
